+++ CHANGELOG.md
@@ -3,10 +3,13 @@
 #### Features
 
 * Your contribution here.
+* [#793](https://github.com/ruby-grape/grape-swagger/pull/793): Features/inheritance and discriminator - [@MaximeRDY](https://github.com/MaximeRDY).
 
 #### Fixes
 
 * Your contribution here.
+* [#798](https://github.com/ruby-grape/grape-swagger/pull/798): Modify full entity name separator - [@GarrettB71](https://github.com/GarrettB71).
+* [#796](https://github.com/ruby-grape/grape-swagger/pull/796): Support grape 1.4.0 - [@thedanielhanke](https://github.com/thedanielhanke).
 
 
 ### 1.1.0 (April 20, 2020)